tweaks-position: Add javascript to handle video position and size
Trying to handle all of the possible view scenarios with CSS alone is
proving difficult, especially with flexbox limitations on vertical
layouts. Instead, switch to using javascript to handle the position and
size. This gives a lot of flexibility to arrange the video windows based
on other elements on the screen.