tweak-position: Move the Change Position button to the left bar
The change-size button is not available when using an audio-only
configuration or if there is no current connection to a WebRTC server.
To make sure this is always available, move the button to the left
navigation bar of the user's camera box.

This has the added benefit of clearing up some of the clutter in the
bottom bar.

Fixes: https://github.com/bekriebel/fvtt-module-webrtc_tweaks/issues/6